MAINT: Rework to nlcg namespace
