+++ CppCore/tests/test_optim_cg.cc
@@ -18,13 +18,14 @@
 #include "xtsci/optimize/linesearch/step_size/geom.hpp"
 #include "xtsci/optimize/linesearch/step_size/golden.hpp"
 
-#include "xtsci/optimize/minimize/cg.hpp"
+#include "xtsci/optimize/minimize/nlcg.hpp"
+#include "xtsci/optimize/nlcg/base.hpp"
 
 #include "xtsci/optimize/trial_functions/quadratic.hpp"
 #include "xtsci/optimize/trial_functions/rosenbrock.hpp"
 
-#include "xtsci/optimize/linesearch/conjugacy/fletcher_reeves.hpp"
-#include "xtsci/optimize/linesearch/conjugacy/polak_ribiere.hpp"
+#include "xtsci/optimize/nlcg/conjugacy/fletcher_reeves.hpp"
+#include "xtsci/optimize/nlcg/conjugacy/polak_ribiere.hpp"
 
 #include <catch2/catch_all.hpp>
 #include <catch2/matchers/catch_matchers_floating_point.hpp>
@@ -32,8 +33,8 @@
 TEST_CASE("ConjugateGradientOptimizer with Backtracking Search") {
   xts::optimize::trial_functions::Rosenbrock<double> rosen;
   xts::optimize::OptimizeControl<double> control;
-  xts::optimize::linesearch::conjugacy::FletcherReeves<double> fletcherreeves;
-  xts::optimize::linesearch::conjugacy::PolakRibiere<double> polakribiere;
+  xts::optimize::nlcg::conjugacy::FletcherReeves<double> fletcherreeves;
+  xts::optimize::nlcg::conjugacy::PolakRibiere<double> polakribiere;
   control.tol = 1e-6;
 
   xt::xarray<double> initial_guess = {-1.3, 1.8};
