+++ CppCore/tiny_cli.cpp
@@ -22,11 +22,11 @@
 #include "xtsci/optimize/linesearch/search_strategy/moore_thuente.hpp"
 #include "xtsci/optimize/linesearch/search_strategy/zoom.hpp"
 
-#include "xtsci/optimize/linesearch/conjugacy/fletcher_reeves.hpp"
-#include "xtsci/optimize/linesearch/conjugacy/hestenes-stiefel.hpp"
-#include "xtsci/optimize/linesearch/conjugacy/hybridized_conj.hpp"
-#include "xtsci/optimize/linesearch/conjugacy/liu_storey.hpp"
-#include "xtsci/optimize/linesearch/conjugacy/polak_ribiere.hpp"
+#include "xtsci/optimize/nlcg/conjugacy/fletcher_reeves.hpp"
+#include "xtsci/optimize/nlcg/conjugacy/hestenes-stiefel.hpp"
+#include "xtsci/optimize/nlcg/conjugacy/hybridized_conj.hpp"
+#include "xtsci/optimize/nlcg/conjugacy/liu_storey.hpp"
+#include "xtsci/optimize/nlcg/conjugacy/polak_ribiere.hpp"
 
 #include "xtsci/optimize/linesearch/step_size/bisect.hpp"
 #include "xtsci/optimize/linesearch/step_size/cubic.hpp"
@@ -105,11 +105,11 @@ int main(int argc, char *argv[]) {
   xts::optimize::linesearch::search_strategy::MooreThuenteLineSearch<double>
       moorethuente(bisectionStep, 1e-3, 0.3);
 
-  xts::optimize::linesearch::conjugacy::FletcherReeves<double> fletcherreeves;
-  xts::optimize::linesearch::conjugacy::PolakRibiere<double> polakribiere;
-  xts::optimize::linesearch::conjugacy::HestenesStiefel<double> hestenesstiefel;
-  xts::optimize::linesearch::conjugacy::LiuStorey<double> liustorey;
-  xts::optimize::linesearch::conjugacy::HybridizedConj<double> hybrid_min(
+  xts::optimize::nlcg::conjugacy::FletcherReeves<double> fletcherreeves;
+  xts::optimize::nlcg::conjugacy::PolakRibiere<double> polakribiere;
+  xts::optimize::nlcg::conjugacy::HestenesStiefel<double> hestenesstiefel;
+  xts::optimize::nlcg::conjugacy::LiuStorey<double> liustorey;
+  xts::optimize::nlcg::conjugacy::HybridizedConj<double> hybrid_min(
       hestenesstiefel, polakribiere,
       [](double a, double b) -> double { return std::min(a, b); });
