+++ CppCore/xtsci/optimize/nlcg/conjugacy/fletcher_reeves.hpp
@@ -9,27 +9,27 @@
 
 #include "xtensor-blas/xlinalg.hpp"
 
-#include "xtsci/optimize/linesearch/base.hpp"
+#include "xtsci/optimize/nlcg/base.hpp"
 
 namespace xts {
 namespace optimize {
-namespace linesearch {
+namespace nlcg {
 namespace conjugacy {
 template <typename ScalarType>
-class HestenesStiefel : public ConjugacyCoefficientStrategy<ScalarType> {
+class FletcherReeves : public ConjugacyCoefficientStrategy<ScalarType> {
 public:
   ScalarType
   computeBeta(const ConjugacyContext<ScalarType> &ctx) const override {
-    auto grad_change = ctx.current_gradient - ctx.previous_gradient;
-    // [NJWS] Equation 5.46
-    return (xt::linalg::dot(ctx.current_gradient, grad_change)() /
-            xt::linalg::dot(grad_change, ctx.previous_direction)());
+    // [NJWS] Equation 5.41a
+    return xt::linalg::dot(ctx.current_gradient, ctx.current_gradient)() /
+           xt::linalg::dot(ctx.previous_gradient, ctx.previous_gradient)();
   }
 
   // References:
   // [NJWS] Nocedal, J., & Wright, S. (2006). Numerical optimization. Springer
 };
+
 } // namespace conjugacy
-} // namespace linesearch
+} // namespace nlcg
 } // namespace optimize
 } // namespace xts
