+++ CppCore/xtsci/optimize/nlcg/conjugacy/hager_zhang.hpp
@@ -9,11 +9,11 @@
 
 #include "xtensor-blas/xlinalg.hpp"
 
-#include "xtsci/optimize/linesearch/base.hpp"
+#include "xtsci/optimize/nlcg/base.hpp"
 
 namespace xts {
 namespace optimize {
-namespace linesearch {
+namespace nlcg {
 namespace conjugacy {
 template <typename ScalarType>
 class HagerZhang : public ConjugacyCoefficientStrategy<ScalarType> {
@@ -42,6 +42,6 @@ public:
 };
 
 } // namespace conjugacy
-} // namespace linesearch
+} // namespace nlcg
 } // namespace optimize
 } // namespace xts
