+++ CppCore/xtsci/optimize/nlcg/conjugacy/hestenes-stiefel.hpp
@@ -9,27 +9,27 @@
 
 #include "xtensor-blas/xlinalg.hpp"
 
-#include "xtsci/optimize/linesearch/base.hpp"
+#include "xtsci/optimize/nlcg/base.hpp"
 
 namespace xts {
 namespace optimize {
-namespace linesearch {
+namespace nlcg {
 namespace conjugacy {
 template <typename ScalarType>
-class PolakRibiere : public ConjugacyCoefficientStrategy<ScalarType> {
+class HestenesStiefel : public ConjugacyCoefficientStrategy<ScalarType> {
 public:
   ScalarType
   computeBeta(const ConjugacyContext<ScalarType> &ctx) const override {
     auto grad_change = ctx.current_gradient - ctx.previous_gradient;
-    // [NJWS] Equation 5.44
-    return xt::linalg::dot(ctx.current_gradient, grad_change)() /
-           xt::linalg::dot(ctx.previous_gradient, ctx.previous_gradient)();
+    // [NJWS] Equation 5.46
+    return (xt::linalg::dot(ctx.current_gradient, grad_change)() /
+            xt::linalg::dot(grad_change, ctx.previous_direction)());
   }
 
   // References:
   // [NJWS] Nocedal, J., & Wright, S. (2006). Numerical optimization. Springer
 };
 } // namespace conjugacy
-} // namespace linesearch
+} // namespace nlcg
 } // namespace optimize
 } // namespace xts
