+++ CppCore/xtsci/optimize/nlcg/conjugacy/liu_storey.hpp
@@ -9,11 +9,11 @@
 
 #include "xtensor-blas/xlinalg.hpp"
 
-#include "xtsci/optimize/linesearch/base.hpp"
+#include "xtsci/optimize/nlcg/base.hpp"
 
 namespace xts {
 namespace optimize {
-namespace linesearch {
+namespace nlcg {
 namespace conjugacy {
 template <typename ScalarType>
 class LiuStorey : public ConjugacyCoefficientStrategy<ScalarType> {
@@ -38,6 +38,6 @@ public:
   // https://doi.org/10.1007/BF00940464.
 };
 } // namespace conjugacy
-} // namespace linesearch
+} // namespace nlcg
 } // namespace optimize
 } // namespace xts
